# backend/common-lib/src/main/java/com/ridedriveahead/common/dto/CancelBookingRequest.java
package com.ridedriveahead.common.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CancelBookingRequest {
    @NotBlank(message = "Cancellation reason is required")
    private String reason;
}