# backend/common-lib/src/main/java/com/ridedriveahead/common/dto/OtpRequest.java
package com.ridedriveahead.common.dto;

import com.ridedriveahead.common.model.enums.UserRole;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class OtpRequest {
    @NotBlank(message = "Phone number is required")
    private String phoneNumber;

    @NotNull(message = "User role is required")
    private UserRole userRole;
}